cleanup init file
